Added Controller method to create User with Profile and Role in one call. To achieve that added User Model object and modified other dependent code.

+++ src/main/java/com/solutionwerk/qb/service/AccountServiceBean.java
@@ -1,7 +1,14 @@
 package com.solutionwerk.qb.service;
 
 import com.solutionwerk.qb.model.Account;
+import com.solutionwerk.qb.model.Profile;
+import com.solutionwerk.qb.model.Role;
+import com.solutionwerk.qb.model.network.User;
 import com.solutionwerk.qb.repository.AccountRepository;
+import com.solutionwerk.qb.repository.ProfileRepository;
+import com.solutionwerk.qb.repository.RoleRepository;
+import com.solutionwerk.qb.util.RequestContext;
+import org.joda.time.DateTime;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -14,7 +21,10 @@ import org.springframework.transaction.annotation.Transactional;
 
 import javax.persistence.EntityExistsException;
 import javax.persistence.NoResultException;
+import java.util.ArrayList;
 import java.util.Collection;
+import java.util.List;
+import java.util.stream.Collectors;
 
 /**
  * The AccountServiceBean encapsulates all business behaviors for operations on
@@ -36,6 +46,12 @@ public class AccountServiceBean implements AccountService {
      */
     @Autowired
     private AccountRepository accountRepository;
+
+    @Autowired
+    private RoleRepository roleRepository;
+
+    @Autowired
+    private ProfileRepository profileRepository;
 
     @Override
     public Collection<Account> findAll() {
@@ -118,7 +134,80 @@ public class AccountServiceBean implements AccountService {
     public void delete(Long id) {
         LOGGER.info("delete id:{}", id);
 
+        if (RequestContext.getUsername() != null && RequestContext.getUsername().equals(accountRepository.findOne(id).getUsername())) {
+            LOGGER.error("Attempted to Delete an Account through which user has logged in.");
+            throw new UnsupportedOperationException("Cannot Delete Logged in User");
+        }
+
         accountRepository.delete(id);
+    }
+
+    @Override
+    @Transactional(
+            propagation = Propagation.REQUIRED,
+            readOnly = false)
+    @CachePut(
+            value = "accounts",
+            key = "#result.id")
+    public Account create(User user) {
+        LOGGER.info("create");
+
+        // Ensure the entity object to be created does NOT exist in the
+        // repository. Prevent the default behavior of save() which will update
+        // an existing entity if the entity matching the supplied id exists.
+        if (user.getAccount().getId() != null) {
+            // Cannot create Account with specified ID value
+            LOGGER.error("Attempted to create an Account, but id attribute was not null.");
+            throw new EntityExistsException("The id attribute must be null to persist a new entity.");
+        }
+
+        Account createdAccount = accountRepository.saveAndFlush(user.getAccount());
+
+        boolean update = false;
+        if (user.getRoles() != null) {
+            List<Role> userRoles = new ArrayList<>(user.getRoles().size());
+            userRoles.addAll(user.getRoles().stream().map(role -> roleRepository.findByCodeAndEffective(role.getCode(), new DateTime())).collect(Collectors.toList()));
+            createdAccount.setRoles(userRoles);
+            update = true;
+        }
+
+        if (user.getProfile() != null) {
+            Profile createdProfile = profileRepository.save(user.getProfile());
+            createdAccount.setProfile(createdProfile);
+            update = true;
+        }
+
+        if (update) {
+            accountRepository.save(createdAccount);
+        }
+
+        return createdAccount;
+    }
+
+    @Override
+    @Transactional(
+            propagation = Propagation.REQUIRED,
+            readOnly = false)
+    @CachePut(
+            value = "accounts",
+            key = "#result.id")
+    public Account update(User user) {
+        LOGGER.info("update id:{}", user.getAccount().getId());
+
+        // Ensure the entity object to be updated exists in the repository to
+        // prevent the default behavior of save() which will persist a new
+        // entity if the entity matching the id does not exist
+        Account accountToUpdate = findOne(user.getAccount().getId());
+        if (accountToUpdate == null) {
+            // Cannot update Account that hasn't been persisted
+            LOGGER.error("Attempted to update a Account, but the entity does not exist.");
+            throw new NoResultException("Requested entity not found.");
+        }
+
+        // TODO: Try to save Account passed in parameter directly. keep the check and ensure other dependent entities like Role and Profiles are intact or updated according to new one.
+        accountToUpdate.setUsername(user.getAccount().getUsername());
+        // TODO: Set Role and Update Pofile values as well
+        return accountRepository.save(accountToUpdate);
     }
 
     @Override
